pyproject.toml
[project]
name = "shop"
version = "0.1.0"
description = ""
authors = [
    {name = "Vlad0n20",email = "vlad.onischuk1234@gmail.com"}
]
readme = "README.md"
requires-python = "3.13.2"
dependencies = [
    "django>=5.1.7,<6.0.0",
    "djangorestframework>=3.15.2,<4.0.0",
    "drf-yasg>=1.21.10,<2.0.0",
    "celery>=5.4.0,<6.0.0",
    "redis>=5.2.1,<6.0.0",
    "pillow>=11.1.0,<12.0.0",
    "psycopg2>=2.9.10,<3.0.0",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "django-filter (>=25.1,<26.0)",
    "faker (>=37.0.2,<38.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "faker-commerce (>=1.0.4,<2.0.0)",
    "pre-commit (>=4.2.0,<5.0.0)",
    "bson (>=0.5.10,<0.6.0)"
]

[tool.poetry.group.dev.dependencies]
django-debug-toolbar = ">=5.1.0,<6.0.0"
django-silk = ">=5.3.2,<6.0.0"
ipython = ">=9.0.2,<10.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
